ci: Fix 'invalid character 'c' in tfplugindocs 0.7
Not sure why this occurs, but it seems only to be an issue if it uses
terraform from PATH (which started in v0.7.0) - if it acquires (the same
version of) terraform itself, it's fine.

But also only in CI - it worked fine (with terraform on PATH) for me locally.

cf. hashicorp/terraform-plugin-docs#127